requirements.txt
fastapi
uvicorn
pydantic
python-lichess==0.9
python-chess